+++ CHANGELOG.md
@@ -78,6 +78,9 @@
 - [#4465](https://github.com/ChainSafe/forest/issues/4665) Add support for the
   `Filecoin.EthNewFilter` RPC method.
 
+- [#4667](https://github.com/ChainSafe/forest/issues/4667) Add support for the
+  `Filecoin.EthNewPendingTransactionFilter` RPC method.
+
 - [#4686](https://github.com/ChainSafe/forest/issues/4686) Add support for the
   `Filecoin.EthAddressToFilecoinAddress` RPC method.